Add a new field, RoundingDecimalField.
This field functions the same as a DecimalField, however it accepts a float as a value. When passed, the float
will be rounded to the proper number of decimal places. This is used for the IndicatorData model, to enforce consistent
decimal behavior. Rounding is done by a call to round() (http://docs.python.org/library/functions.html#round)